+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
+- [tracing] fix: Add manual `DOMStringList` typing (#2718)
 
 ## 5.19.0